backend/requirements.txt
Django
psycopg2-binary
daphne
requests
get-docker-secret
djangorestframework
djangorestframework-simplejwt[crypto]
drf-yasg
faker
pytest
pytest-django
pytest-faker
pytest-asyncio
pytest-cov
Twisted[tls,http2]